hub: implementing haproxy agent-check TCP port for draining a hub

+++ src/smc-hub/healthchecks.ts
@@ -8,23 +8,20 @@
 import * as debug from "debug";
 const L = debug("hub:healthcheck");
 import { Router } from "express";
+import { createServer } from "net";
 import { isFloat } from "validator";
 import { seconds2hms } from "../smc-util/misc";
 const { database_is_working } = require("./hub_register");
 import { PostgreSQL } from "./postgres/types";
 
-interface Opts {
-  router: Router;
-  db: PostgreSQL;
-}
-
 // self termination is only activated, if there is a COCALC_HUB_SELF_TERMINATE environment variable
-// it's value is an interval in hours, minimum and maximum, for how long it should live.
-// e.g. "24,48" for between 1 and 2 days.
+// it's value is an interval in hours, minimum and maximum, for how long it should be alive
+// and a drain period in minutes at the end.
+// e.g. "24,48,15" for an uptime between 1 and 2 days and 15 minutes of draining
 function init_self_terminate(): {
   startup: number;
   shutdown?: number; // when to shutdown (causes a failed health check)
-  dead?: number; // when to return not-alive, causes a proxy server to no longer send traffic
+  drain?: number; // when to start draining, causes a proxy server to no longer send traffic
 } {
   const startup = Date.now();
   const conf = process.env.COCALC_HUB_SELF_TERMINATE;
@@ -32,32 +29,72 @@ function init_self_terminate(): {
     L("COCALC_HUB_SELF_TERMINATE not set, hence no self-termination");
     return { startup };
   }
-  const [from_str, to_str] = conf.trim().split(",");
+  const [from_str, to_str, drain_str] = conf.trim().split(",");
   if (!isFloat(from_str, { gt: 0 }))
-    throw Error("COCALC_HUB_SELF_TERMINATE/from not a positive float");
+    throw new Error("COCALC_HUB_SELF_TERMINATE/from not a positive float");
   if (!isFloat(to_str, { gt: 0 }))
-    throw Error("COCALC_HUB_SELF_TERMINATE/to not a positive float");
+    throw new Error("COCALC_HUB_SELF_TERMINATE/to not a positive float");
+  if (!isFloat(drain_str, { gt: 0 }))
+    throw new Error("COCALC_HUB_SELF_TERMINATE/drain not a positive float");
   const from = parseFloat(from_str);
   const to = parseFloat(to_str);
+  const drain_h = parseFloat(drain_str) / 60; // minutes to hours
   if (from > to)
     throw Error(
-      "COCALC_HUB_SELF_TERMINATE from must be smaller than to, e.g. '24,48'"
+      "COCALC_HUB_SELF_TERMINATE 'from' must be smaller than 'to', e.g. '24,48,15'"
     );
   const uptime = Math.random() * (to - from); // hours
   const hours2ms = 1000 * 60 * 60;
   const shutdown = startup + uptime * hours2ms;
-  // controlled shutdown: send out being "dead" 10 minutes or 10% before the actual shutdown
-  const dead_period = Math.min(1 / 6, 0.1 * uptime); // hours
-  const dead = shutdown - dead_period * hours2ms;
+  const drain = shutdown - drain_h * hours2ms;
+  if (startup > drain) {
+    throw new Error(
+      `COCALC_HUB_SELF_TERMINATE: startup must be smaller than drain – ${startup}>${drain}`
+    );
+  }
   L(
-    `init_self_terminate: startup=${startup} dead=${dead} shutdown=${shutdown} uptime=${seconds2hms(
-      hours2ms * uptime
-    )}`
+    `init_self_terminate: startup=${startup} drain=${drain} shutdown=${shutdown} uptime=${seconds2hms(
+      (hours2ms * uptime) / 1000
+    )} draintime=${seconds2hms((drain_h * hours2ms) / 1000)}`
   );
-  return { startup, shutdown, dead };
+  return { startup, shutdown, drain };
 }
 
-const { startup, shutdown, dead } = init_self_terminate();
+const { startup, shutdown, drain } = init_self_terminate();
+
+let agent_port = 0;
+export function set_agent_port(port: number) {
+  agent_port = port;
+}
+
+let agent_check_server: any;
+
+// https://cbonte.github.io/haproxy-dconv/2.0/configuration.html#5.2-agent-check
+// for development, set the env var in your startup script or terminal init file
+// export COCALC_HUB_SELF_TERMINATE=.1,.2,1
+// and then query it like that
+// $ telnet 0.0.0.0 $(cat $SMC_ROOT/dev/project/ports/agent-port)
+function setup_agent_check() {
+  if (agent_port == 0 || drain == null) {
+    L("setup_agent_check: agent_port not set, no agent checks");
+    return;
+  }
+
+  // TODO this could also return a "weight" for this server, based on load values
+  agent_check_server = createServer((c) => {
+    let msg = Date.now() < drain ? "ready" : "drain";
+    c.write(msg + "\r\n");
+    c.destroy();
+  });
+
+  agent_check_server.listen(agent_port, "0.0.0.0");
+  L(`setup_agent_check: setup on port ${agent_port}`);
+}
+
+interface Opts {
+  router: Router;
+  db: PostgreSQL;
+}
 
 interface Check {
   status: string;
@@ -66,6 +103,7 @@ interface Check {
 
 export async function setup_healthchecks(opts: Opts): Promise<void> {
   const { router, db } = opts;
+  setup_agent_check();
 
   // used by HAPROXY for testing that this hub is OK to receive traffic
   router.get("/alive", (_, res) => {
@@ -76,7 +114,7 @@ export async function setup_healthchecks(opts: Opts): Promise<void> {
       // this will stop haproxy from routing traffic to us
       // until db connection starts working again.
       msg = "alive: NO – database not working";
-    } else if (dead != null && Date.now() > dead) {
+    } else if (shutdown != null && Date.now() > shutdown) {
       msg = "alive: NO – shutdown initiated";
     } else {
       is_dead = false;
@@ -100,14 +138,18 @@ export async function setup_healthchecks(opts: Opts): Promise<void> {
   function check_uptime(): Check {
     const now = Date.now();
     const uptime = seconds2hms((now - startup) / 1000);
-    if (shutdown != null) {
+    if (shutdown != null && drain != null) {
       if (now >= shutdown) {
         const msg = `uptime ${uptime} – expired, terminating now`;
         L(msg);
         return { status: msg, abort: true };
       } else {
         const until = seconds2hms((shutdown - now) / 1000);
-        const msg = `uptime ${uptime} – terminating in ${until}`;
+        const drain_str =
+          drain > now
+            ? `draining in ${seconds2hms((drain - now) / 1000)}`
+            : "draining now";
+        const msg = `uptime ${uptime} – ${drain_str} – terminating in ${until}`;
         L(msg);
         return { status: msg };
       }
